+++ ideal_weight.rb
@@ -3,10 +3,10 @@ p "Введите своё имя:"
 name = gets.chomp
 p "Спасибо, #{name}!"
 p "Теперь введите Ваш рост:"
-height = gets.chomp.to_i
-ideal_weight = (height.to_i - 110) * 1.15
+height = gets.chomp.to_f
+ideal_weight = (height.to_f - 110) * 1.15
 
-if ideal_weight<0
+if ideal_weight < 0
 	p "Ваш вес уже оптимальный"
 else
 	p "Идеальный вес для роста #{height}см, составляет #{ideal_weight} кг"
